refactor: standardize import architecture and eliminate barrel exports
