+++ pyproject.toml
@@ -57,8 +57,10 @@ select = [
     "ISC", # flake8-implicit-str-concat
     "RUF", # Ruff-specific rules
     "PL",  # Pylint
-    "PERF",# Perflint
+    "TID252", # flake8-tidy-imports
+    "PYI024", # collections-named-tuple
 ]
+
 ignore = [
     "UP043", # Unnecessary default type arguments
     "PLR2004", # Magic value used in comparison
@@ -66,6 +68,9 @@ ignore = [
     "PLR0913", # Too many arguments
     "PLR0915", # Too many statements
 ]
+
+[tool.ruff.lint.flake8-tidy-imports]
+ban-relative-imports = "all"
 
 [tool.ruff.lint.mccabe]
 max-complexity = 15
