+++ src/python/claude_statusline/claude_statusline/segments/__init__.py
@@ -1,31 +1,1 @@
-from claude_statusline.models import SegmentGenerationResult
-from claude_statusline.segments.claude import (
-    format_context_usage,
-    format_cost,
-    format_model_info,
-    format_session_info,
-)
-from claude_statusline.segments.constants import (
-    DIVIDER_BAR,
-    DIVIDER_DOT,
-)
-from claude_statusline.segments.git import format_git_full
-from claude_statusline.segments.workspace import (
-    format_directory,
-    format_obsidian_vault,
-    shorten_path,
-)
 
-__all__ = [
-    "DIVIDER_BAR",
-    "DIVIDER_DOT",
-    "SegmentGenerationResult",
-    "format_context_usage",
-    "format_cost",
-    "format_directory",
-    "format_git_full",
-    "format_model_info",
-    "format_obsidian_vault",
-    "format_session_info",
-    "shorten_path",
-]
