+++ src/python/termtext/src/termtext/__init__.py
@@ -1,25 +1,1 @@
-from termtext.attribute import BG, FG, SGR, Attribute, Link
-from termtext.color import Color, Color256, NamedColor, RGBColor
-from termtext.config import DEFAULT_CONFIG, Config
-from termtext.segment import EscapedText, Segment
-from termtext.style import Style
-from termtext.term_text import TermText
 
-__all__ = [
-    "Attribute",
-    "BG",
-    "Color",
-    "Color256",
-    "Config",
-    "DEFAULT_CONFIG",
-    "EscapedText",
-    "FG",
-    "Link",
-    "NamedColor",
-    "RGBColor",
-    "SGR",
-    "Segment",
-    "Style",
-    "TermText",
-]
-# ruff: noqa: RUF022
